Updated the nf pipeline and added appropriate python scripts for inital folding operations
