+++ setup.py
@@ -69,16 +69,16 @@ setup(name="mwa_search",
                'plotting/plot_obs_pulsar.py',
                'plotting/position_sn_heatmap_fwhm.py',
                'dpp/binfinder.py', 'dpp/stokes_fold.py', 'dpp/plotting_toolkit.py',
-               'dpp/yaml_helper.py', 'dpp/pipe_helper.py', 'dpp/prepfold_launch.py',
-               'dpp/pulsar_obs_helper.py', 'dpp/dpp_check_args.py', 'dpp/misc_helper.py', 'dpp/submit_move.py',
+               'dpp/yaml_helper.py', 'dpp/prepfold_cmd_make.py',
+               'dpp/pulsar_obs_helper.py', 'dpp/dpp_check_args.py', 'dpp/misc_helper.py', 'dpp/upload_and_move.py',
                'version.py',
                'nextflow/beamform.nf', 'nextflow/beamform_module.nf',
-               'nextflow/beamform_fov_sources.nf', 'nextflow/nextflow.config',
+                'nextflow/nextflow.config',
                'nextflow/candidate_TOAs.nf', 'nextflow/find_candidate_position.nf',
                'nextflow/classifier.nf', 'nextflow/classifier_module.nf',
                'nextflow/pdmp.nf', 'nextflow/dspsr_module.nf',
                'nextflow/pulsar_search.nf', 'nextflow/pulsar_search_module.nf',
-               'nextflow/mwa_search_pipeline.nf'],
+               'nextflow/mwa_search_pipeline.nf', 'nextflow/data_processing_pipeline.nf'],
       #data_files=[('AegeanTools', [os.path.join(data_dir, 'MOC.fits')]) ],
       setup_requires=['pytest-runner'],
       tests_require=['pytest']  # , 'nose']
